Score, Highscore, Highscore file, ...
Window now has a label for showing the score and highscore
Implemented code for saving and loading all highscores from a file.
This file has a highscore per line.
When the game starts, we load the file and set highscore as the highest
value.
When you lose, we save to the end of the file a new line with the
current highscore.
Class App was rebuild, now constains new methods (create_grid, update,
mainloop, destroy, save_highscore)
Removed window argument for init method for Grid class.
Checking condition for the snake eating food has been moved to a
different function.
Score increment by 10 when the snake eats the food.