Save Highscore, Fix Collision, Movement mechanism
main.py: all_highscores is a list of all highscores stored in the file,
we will only save a new highscore to that file if it's not included in
that list.
Grid.py: modified the method to obtaining the maximum highscore stored
in the file. The output is the same, what changes is the code. In show
method we now draw every cell, so we can have control over the color of
every cell of the grid. The code to show the snake body has been moved
to the Snake Class, which makes more sense on a structured programming
perspective. And fixed a bug on collision between the snake and the
walls
Snake.py: The movement mechanism is different now. Instead of changing
the velocity whenever a key is pressed, now everytime a key is pressed a
movement is added to a list of moves (moves_buffer), then when we move
we remove the first element to update the velocity of the snake and
update it's position. I also added a mapping between the arrows and a
velocity, so now anyone can play with the wasd layout or the arrows
layout.
Vector.py: added some more methods to the Vector class like mul (scalar product of a
vector), neg (vector * -1), scale method (scalar product of a vector in
situ). Function lerp, does a linear interpolation between 2 vectores.