Bolus, Snake width
Grid.py: whenever the snake eats the food, we set the snake bolus
property to 0.

Snake.py: class Snake has a new property bolus, which it is the index of
the snake body where is the bolus (food)
Snake.py: now the snake body is a scale of the width, the width of each
segment of the snake's body doesn't need to be the same width of the
cell grid.