Robot Base implementation
main.py: App class now has a property which determines if the robot is
playing (is_robot_playing). The keys don't work if the robot is playing.
The robot takes over the human player. added a show method, I removed
from the Grid class, because it makes more sense that the Grid class
doesn't have to deal with the canvas.

Grid.py: Removed show function, grid update only deals with it's own
matrix of cell locations and the snake. The update method only moves the
snake, check if there are collisions and check if the snake is eating
the food.

Robot.py: New file added which will contain a Robot class that will play
the game.