Improved Snake "AI" strategy
main.py: Now the robot play after the grid has been updated.

Robot.py: The strategy is picking a valid move (up, left, right or down)
that is the closest to the food. A valid move is a move that will not
collide with the wall neither with it's own body. There is some problems
with this strategy, naming for example sometimes it won't have an
alternative to chose a move that will eventually lead to the snake be
trapped in a closed area with it's own body.