A* path finding implemented
main.py: Now we pass a dictionary to play method of Robot class. This
dictionary contains the came from path from a* star path finding
algorithm. This dictionary map a position to the position it came from.

Grid.py: The A_star, reconstruct_path, h functions have been removed
from Robot.py and moved to Grid.py. The came_from_path is generated in
the Grid Class in initialization and after a food has been generated

Robot.py: in play method of Robot class, the previous strategy (moving
to a neighbor that is closest to the food) with the new strategy
following the next position obtained by the a-star path finding
algorithm