Added modularity to path generator
Grid.py: Now generate_path method from Grid class has the possibility to
feed starting position, the goal position and the matrix with the walls.
All these arguments are optional, if none are given then it will work
with the default start = food, goal = snake_head, the walls are the
snake body segments excluding the head. The function also return the
dictionary that maps a location to where it came from.