Grid cell width whole division bug
The problem was that when the number of rows/cols didn't divide the
canvas width, it showed black region at the right and bottom part of the
canvas. This was due to a whole division when we calculate the width of
the cell of the grid. Now the width and height are obtained through a
float division.

main.py: dw, dh are now declared as floats.

Snake.py: show method from class Snake, now receives dw (cell's width)
and dh (cell's height) as float.