Choose strategy, More strategies to choose
main.py: The default strategy is A* path finding algorithm

Grid.py: generate_path return None, but changes in situ cmae_from_path

Robot.py: Removed Hamiltonian cycle (not yet implemented), added 2
strategies Random Walk, and Cycle.

play method will play according to the strategy it is chosen.

In Random Walk, the robot chooses a random neighbor that is inside the
grid and is not a segment of the snake body.

The cycle performs a cycle through the grid going through every point.
Is not very efficient, it takes a lot of steps to eat the food, but It
will win 100% of the time.