Optimizing show method
main.py: Instead of redrawing everything every frame. The grid is drawn
on the canvas once when the grid is created (create_grid), and every
element that is drawn a frame is deleted after updating the canvas. show
method of Snake class has been moved to the App class.

Snake.py: show method has been moved to the App class.